Add a core.telemetry module for configuring otel
