+++ pixl_export/src/pixl_export/main.py
@@ -23,7 +23,7 @@ from pathlib import Path
 from typing import Annotated
 
 from core.exports import ParquetExport
-from core.logging import configure_logging
+from core.telemetry import configure_logging
 from core.rest_api.router import router
 from core.uploader import get_uploader
 from decouple import config  # type: ignore [import-untyped]
