+++ pixl_imaging/src/pixl_imaging/main.py
@@ -18,7 +18,7 @@ from __future__ import annotations
 import asyncio
 import importlib.metadata
 
-from core.logging import configure_logging
+from core.telemetry import configure_logging
 from core.patient_queue.subscriber import PixlConsumer
 from core.rest_api.router import router, state
 from decouple import config
